src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class StatsServiceTest {
    @Test

    public void totalSum() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};

        long actual = service.total(sales);
        long expected = 180;

        Assertions.assertEquals(expected, actual);

    }

    @Test

    public void AverageSum() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};

        long actual = service.average(sales);
        long expected = 15;

        Assertions.assertEquals(expected, actual);

    }

    @Test

    public void maxMonth() {

        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};

        int actual = service.maxMonth(sales);
        int expected = 7;

        Assertions.assertEquals(expected, actual);
    }

    @Test

    public void minMonth() {

        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};

        int actual = service.minMonth(sales);
        int expected = 8;

        Assertions.assertEquals(expected, actual);
    }


    @Test

    public void underAverage() {

        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};

        long actual = service.underAverage(sales);
        int expected = 5;

        Assertions.assertEquals(expected, actual);

    }

    @Test

    public void overAverage() {

        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};

        long actual = service.overAverage(sales);
        int expected = 5;

        Assertions.assertEquals(expected, actual);

    }

}

